Update Main.py
It will be good if the player can see all the written texts clearly. Previous color-coding wasn't good enough, so I decided to make a little change to how the program actually displays the high-score, and current-score text... I think white will be quite better, so I have changed the color to white.